# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"User": {
				"properties": {
					"Id": {
						"type": "number",
						"format": "double"
					},
					"FirstName": {
						"type": "string"
					},
					"LastName": {
						"type": "string"
					},
					"Email": {
						"type": "string"
					}
				},
				"required": [
					"FirstName",
					"LastName",
					"Email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PagedData_User_": {
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					},
					"pageSize": {
						"type": "number",
						"format": "double"
					},
					"pageOffset": {
						"type": "number",
						"format": "double"
					},
					"count": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"items",
					"pageSize",
					"pageOffset",
					"count"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateUserCommandResponse": {
				"properties": {
					"Id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"Id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateUserCommand": {
				"properties": {
					"FirstName": {
						"type": "string"
					},
					"LastName": {
						"type": "string"
					},
					"Email": {
						"type": "string"
					},
					"Roles": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"Password": {
						"type": "string"
					}
				},
				"required": [
					"FirstName",
					"LastName",
					"Email",
					"Roles",
					"Password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthClaims": {
				"properties": {
					"FirstName": {
						"type": "string"
					},
					"LastName": {
						"type": "string"
					},
					"Email": {
						"type": "string"
					},
					"Roles": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"FirstName",
					"LastName",
					"Email",
					"Roles"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthToken": {
				"properties": {
					"token": {
						"type": "string"
					}
				},
				"required": [
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthenticationLoginRequest": {
				"properties": {
					"userName": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"userName",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"features": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "nodejs-express-api",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Paul Salmon"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/users": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PagedData_User_"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"features": [
							"users.list"
						]
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateUserCommandResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"features": [
							"users.add"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserCommand"
							}
						}
					}
				}
			}
		},
		"/users/current": {
			"get": {
				"operationId": "GetCurrent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthClaims"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": []
			}
		},
		"/auth": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthToken"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AuthenticationLoginRequest"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:5000/api/v1"
		}
	]
}